Move the dynamic port selection to its own util, and unify develop and serve-build on it
